Split server definition into smaller files
