+++ lib/types/server/state.d.ts
@@ -1,0 +1,138 @@
+
+import { SealOptions, SealOptionsSub } from '@hapi/iron';
+
+import { Request } from '../request';
+import { Utils } from '../utils';
+
+export type SameSitePolicy = false | 'None' | 'Lax' | 'Strict';
+
+/**
+ * Optional cookie settings
+ * [See docs](https://github.com/hapijs/hapi/blob/master/API.md#-serverstatename-options)
+ */
+export interface ServerStateCookieOptions {
+    /** time-to-live in milliseconds. Defaults to null (session time-life - cookies are deleted when the browser is closed). */
+    ttl?: number | null | undefined;
+    /** sets the 'Secure' flag. Defaults to true. */
+    isSecure?: boolean | undefined;
+    /** sets the 'HttpOnly' flag. Defaults to true. */
+    isHttpOnly?: boolean | undefined;
+    /**
+     * sets the 'SameSite' flag. The value must be one of:
+     *  * false - no flag.
+     *  * 'Strict' - sets the value to 'Strict' (this is the default value).
+     *  * 'Lax' - sets the value to 'Lax'.
+     */
+    isSameSite?: SameSitePolicy | undefined;
+    /** the path scope. Defaults to null (no path). */
+    path?: string | null | undefined;
+    /** the domain scope. Defaults to null (no domain). */
+    domain?: string | null | undefined;
+
+    /**
+     * if present and the cookie was not received from the client or explicitly set by the route handler, the
+     * cookie is automatically added to the response with the provided value. The value can be
+     * a function with signature async function(request) where:
+     */
+    autoValue?(request: Request): void;
+
+    /**
+     * encoding performs on the provided value before serialization. Options are:
+     *  * 'none' - no encoding. When used, the cookie value must be a string. This is the default value.
+     *  * 'base64' - string value is encoded using Base64.
+     *  * 'base64json' - object value is JSON-stringified then encoded using Base64.
+     *  * 'form' - object value is encoded using the x-www-form-urlencoded method.
+     *  * 'iron' - Encrypts and sign the value using iron.
+     */
+    encoding?: 'none' | 'base64' | 'base64json' | 'form' | 'iron' | undefined;
+    /**
+     * an object used to calculate an HMAC for cookie integrity validation. This does not provide privacy, only a mean
+     * to verify that the cookie value was generated by the server. Redundant when 'iron' encoding is used. Options are:
+     *  * integrity - algorithm options. Defaults to require('@hapi/iron').defaults.integrity.
+     *  * password - password used for HMAC key generation (must be at least 32 characters long).
+     */
+    sign?: {
+        integrity?: SealOptionsSub | undefined;
+        password: string;
+    } | undefined;
+    /** password used for 'iron' encoding (must be at least 32 characters long). */
+    password?: string | undefined;
+    /** options for 'iron' encoding. Defaults to require('@hapi/iron').defaults. */
+    iron?: SealOptions | undefined;
+    /** if true, errors are ignored and treated as missing cookies. */
+    ignoreErrors?: boolean | undefined;
+    /** if true, automatically instruct the client to remove invalid cookies. Defaults to false. */
+    clearInvalid?: boolean | undefined;
+    /** if false, allows any cookie value including values in violation of RFC 6265. Defaults to true. */
+    strictHeader?: boolean | undefined;
+    /** used by proxy plugins (e.g. h2o2). */
+    passThrough?: any;
+}
+
+/**
+ * A single object or an array of object where each contains:
+ * * name - the cookie name.
+ * * value - the cookie value.
+ * * options - cookie configuration to override the server settings.
+ */
+export interface ServerStateFormat {
+    name: string;
+    value: string;
+    options: ServerStateCookieOptions;
+}
+
+/**
+ * For context [See docs](https://github.com/hapijs/hapi/blob/master/API.md#-serverstatename-options)
+ * For context [See docs](https://github.com/hapijs/hapi/blob/master/API.md#-serveroptionsstate)
+ */
+export interface ServerState {
+    /**
+     * The server cookies manager.
+     * Access: read only and statehood public interface.
+     */
+    readonly states: object;
+
+    /**
+     * The server cookies manager settings. The settings are based on the values configured in [server.options.state](https://github.com/hapijs/hapi/blob/master/API.md#server.options.state).
+     */
+    readonly settings: ServerStateCookieOptions;
+
+    /**
+     * An object containing the configuration of each cookie added via [server.state()](https://github.com/hapijs/hapi/blob/master/API.md#server.state()) where each key is the
+     * cookie name and value is the configuration object.
+     */
+    readonly cookies: {
+        [key: string]: ServerStateCookieOptions;
+    };
+
+    /**
+     * An array containing the names of all configured cookies.
+     */
+    readonly names: string[];
+
+    /**
+     * Same as calling [server.state()](https://github.com/hapijs/hapi/blob/master/API.md#server.state()).
+     */
+    add(name: string, options?: ServerStateCookieOptions): void;
+
+    /**
+     * Formats an HTTP 'Set-Cookie' header based on the server.options.state where:
+     * @param cookies - a single object or an array of object where each contains:
+     * * name - the cookie name.
+     * * value - the cookie value.
+     * * options - cookie configuration to override the server settings.
+     * @return Return value: a header string.
+     * Note that this utility uses the server configuration but does not change the server state. It is provided for manual cookie formatting (e.g. when headers are set manually).
+     * [See docs](https://github.com/hapijs/hapi/blob/master/API.md#-async-serverstatesformatcookies)
+     */
+    format(cookies: ServerStateFormat | ServerStateFormat[]): Promise<string>;
+
+    /**
+     * Parses an HTTP 'Cookies' header based on the server.options.state where:
+     * @param header - the HTTP header.
+     * @return Return value: an object where each key is a cookie name and value is the parsed cookie.
+     * Note that this utility uses the server configuration but does not change the server state. It is provided for manual cookie parsing (e.g. when server parsing is disabled).
+     * [See docs](https://github.com/hapijs/hapi/blob/master/API.md#-async-serverstatesparseheader)
+     */
+    parse(header: string): Promise<Utils.Dictionary<string>>;
+}
